+++ src/test/java/spoon/test/visibility/VisibilityTest.java
@@ -32,11 +32,11 @@ import static spoon.testing.utils.ModelUtils.canBeBuilt;
 public class VisibilityTest {
     @Test
     public void testMethodeWithNonAccessibleTypeArgument() throws Exception {
-        Factory f = build(spoon.test.visibility.MethodeWithNonAccessibleTypeArgument.class,
+        Factory f = build(spoon.test.visibility.testclasses.MethodeWithNonAccessibleTypeArgument.class,
                 spoon.test.visibility.packageprotected.AccessibleClassFromNonAccessibleInterf.class,
                 Class.forName("spoon.test.visibility.packageprotected.NonAccessibleInterf")
                 );
-        CtClass<?> type = f.Class().get(spoon.test.visibility.MethodeWithNonAccessibleTypeArgument.class);
+        CtClass<?> type = f.Class().get(spoon.test.visibility.testclasses.MethodeWithNonAccessibleTypeArgument.class);
         assertEquals("MethodeWithNonAccessibleTypeArgument", type.getSimpleName());
         CtMethod<?> m = type.getMethodsByName("method").get(0);
         assertEquals(
